load/requirements.txt
pandas==2.0.3
protobuf==4.23.4
psycopg2-binary
PyYAML==6.0.1
